Simplificación del proyecto

+++ src/main/java/com/matias/springjwt/security/services/RefreshTokenService.java
@@ -2,6 +2,7 @@ package com.matias.springjwt.security.services;
 
 import com.matias.springjwt.exception.TokenRefreshException;
 import com.matias.springjwt.model.RefreshTokenEntity;
+import com.matias.springjwt.model.UserEntity;
 import com.matias.springjwt.repository.IRefreshTokenRepository;
 import com.matias.springjwt.repository.IUserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,44 +10,52 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.EntityNotFoundException;
 import java.time.Instant;
 import java.util.Optional;
 import java.util.UUID;
 
 @Service
 public class RefreshTokenService {
-    @Value("${springjwt.app.jwtRefreshExpirationMs}")
-    private Long refreshTokenDurationMs;
 
-    @Autowired
-    private IRefreshTokenRepository refreshTokenRepository;
+	@Value("${springjwt.app.jwtRefreshExpirationMs}")
+	private Long refreshTokenDurationMs;
 
-    @Autowired
-    private IUserRepository userRepository;
+	@Autowired
+	private IRefreshTokenRepository refreshTokenRepository;
 
-    public Optional<RefreshTokenEntity> findByToken(String token) {
-        return refreshTokenRepository.findByToken(token);
-    }
+	@Autowired
+	private IUserRepository userRepository;
 
-    public RefreshTokenEntity createRefreshToken(Long userId) {
-        RefreshTokenEntity refreshToken = new RefreshTokenEntity();
-        refreshToken.setUser(userRepository.findById(userId).get());
-        refreshToken.setExpiryDate(Instant.now().plusMillis(refreshTokenDurationMs));
-        refreshToken.setToken(UUID.randomUUID().toString());
-        refreshToken = refreshTokenRepository.save(refreshToken);
-        return refreshToken;
-    }
+	public Optional<RefreshTokenEntity> findByToken(String token) {
+		return refreshTokenRepository.findByToken(token);
+	}
 
-    public RefreshTokenEntity verifyExpiration(RefreshTokenEntity token) {
-        if (token.getExpiryDate().compareTo(Instant.now()) < 0) {
-            refreshTokenRepository.delete(token);
-            throw new TokenRefreshException(token.getToken(), "Refresh token was expired. Please make a new signin request");
-        }
-        return token;
-    }
+	public RefreshTokenEntity createRefreshToken(Long userId) {
+		RefreshTokenEntity refreshToken = new RefreshTokenEntity();
+		refreshToken.setUser(userRepository.findById(userId).get());
+		refreshToken.setExpiryDate(Instant.now().plusMillis(refreshTokenDurationMs));
+		refreshToken.setToken(UUID.randomUUID().toString());
+		refreshToken = refreshTokenRepository.save(refreshToken);
+		return refreshToken;
+	}
 
-    @Transactional
-    public int deleteByUserId(Long userId) {
-        return refreshTokenRepository.deleteByUser(userRepository.findById(userId).get());
-    }
+	public RefreshTokenEntity verifyExpiration(RefreshTokenEntity token) {
+		if (token.getExpiryDate().compareTo(Instant.now()) < 0) {
+			refreshTokenRepository.delete(token);
+			throw new TokenRefreshException(token.getToken(),
+					"Refresh token was expired. Please make a new signin request");
+		}
+		return token;
+	}
+
+	@Transactional
+	public int deleteByUserId(Long userId) {
+		Optional<UserEntity> optionalUserEntity = userRepository.findById(userId);
+		if (optionalUserEntity.isEmpty()) {
+			throw new EntityNotFoundException("Entity not found");
+		}
+		return refreshTokenRepository.deleteByUser(userRepository.findById(userId).get());
+	}
+
 }
